Way to display "add flag" component must be fixed
The current way flags are being displayed are through a "v-if="showFlagModal === student.email"", when it should probably be set up through a boolean. The issue comes in when you try to close the "add/remove flag" window: The method set up passes in ' ' to break the condition and remove the flag modal through this. Because of this, the student table gets reset to the first page, and the flag modal from the page you opened it on doesn't go away.