many changes
introduced a CourseTable that emulates studentTable wtihout flags, for use wtih courses. CourseList uses this table

changed sortButton to be dynamic by allowing the array doing the sorting to be a prop (which is why student list is also changed). SortArray interface currently unused but maybe enforced later

introduced a getCourses function in the stores with a currently unused courseList

what needs to be done:
more sorting
possibly remove flags from courses, or reuse them for a separate purpose..? this requires making ChangeFlag dynamic